+++ src/gateway.js
@@ -16,12 +16,13 @@ let server;
  *            - manifest:       The manifest to add apps from.
  *            - port:           The port to use.
  *            - publishEvent:   Function that publishes an event across all containers (via RabbitMQ).
+ *            - mainApi:        The main API.
  *
  */
 const start = (settings = {}) => {
     return new Promise((resolve, reject) => {
       Promise.coroutine(function*() {
-        const { apps, manifest, publishEvent } = settings;
+        const { apps, manifest, publishEvent, mainApi } = settings;
         const port = settings.port || DEFAULT_GATEWAY_PORT;
         const middleware = express().use(bodyParser.json());
 
@@ -43,7 +44,7 @@ const start = (settings = {}) => {
         }
 
         // Routes.
-        if (apiRoute) { api({ apiRoute, apps, middleware, manifest, publishEvent }); }
+        if (apiRoute) { api({ apiRoute, apps, middleware, manifest, mainApi }); }
         gatewayRouter(apps, middleware);
 
         // Listen on the desired port.