+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Manifest path within API status.
 - Storing `gatewayPort` on the main API, and remembering it on a manifest restart.
 - `restart()` method to the main API.
+- API: `/api/restart` method.
 #### Changed
 - API webhook route changed from "api/repo" => "api/github".
 - RabbitMQ can only be specified as an environment variable (no longer declared within `manifest.yml`).